pyproject.toml
[project]
name = "ensayo"
version = "0.1.0"
description = "Generator for educational company simulation sites — scaffolds content and builds vanilla HTML sites from markdown and YAML configuration."
readme = "README.md"
license = "MIT"
requires-python = ">=3.11"
authors = [
    { name = "Michael Borck", email = "michael@borck.id.au" },
]
keywords = ["education", "simulation", "generator", "html", "chatbot"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Education",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Education",
    "Topic :: Text Processing :: Markup :: HTML",
]

dependencies = [
    "click>=8.1",
    "jinja2>=3.1",
    "markdown>=3.5",
    "pyyaml>=6.0",
    "rich>=13.0",
]

[project.optional-dependencies]
content = [
    "anthropic>=0.40",
]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "ruff>=0.8",
    "mypy>=1.13",
    "types-PyYAML>=6.0",
]

[project.scripts]
ensayo = "ensayo.cli:main"

[project.urls]
Homepage = "https://github.com/michael-borck/ensayo"
Repository = "https://github.com/michael-borck/ensayo"
Issues = "https://github.com/michael-borck/ensayo/issues"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/ensayo"]

[tool.ruff]
target-version = "py311"
src = ["src"]
line-length = 99

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "B", "SIM", "RUF"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true
mypy_path = "src"

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --tb=short"
